Allow user to select font used for menus, labels, etc.
Add an extra row to the Prefs Font dialog to allow selection of a font for all widgets
except those that already have a specific font. Also, allow the system default to be
used via a checkbox.

Due to Perl/Tk and/or platform limitations, the top menu bar font is not affected
by this commit. It may be possible to rewrite the way the top menu bar is done
to use an alternative method that would be affected by the global font setting.
If possible, this will be in a second commit.

+++ src/lib/Guiguts/MenuStructure.pm
@@ -13,7 +13,9 @@ BEGIN {
 # rebuild every time it is modified
 sub menurebuild {
 
-    $::menubar->delete( 0, 'last' );    # Delete any current menus
+    return unless Tk::Exists($::menubar);    # Abort calls made before menubar is created
+
+    $::menubar->delete( 0, 'last' );         # Delete any current menus
 
     menu_toplevel( '~File',        &menu_file );
     menu_toplevel( '~Edit',        &menu_edit );
